Rename hash_dict.py to Hashing.py
